+++ src/GlobalPayments.Api/Builders/ManagementBuilder.cs
@@ -39,6 +39,7 @@ namespace GlobalPayments.Api.Builders {
         internal string Description { get; set; }
         internal IEnumerable<DisputeDocument> DisputeDocuments { get; set; }
         internal string DisputeId { get; set; }
+        internal string DynamicDescriptor { get; set; }
         internal decimal? Gratuity { get; set; }
         internal string IdempotencyKey { get; set;  }
         internal string InvoiceNumber { get; set; }
@@ -212,6 +213,20 @@ namespace GlobalPayments.Api.Builders {
         /// <returns>ManagementBuilder</returns>
         public ManagementBuilder WithDisputeId(string value) {
             DisputeId = value;
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the transaction's dynamic descriptor.
+        /// </summary>
+        /// <remarks>
+        /// This value is sent during the authorization process and is displayed
+        /// in the consumer's account.
+        /// </remarks>
+        /// <param name="value">The dynamic descriptor</param>
+        /// <returns>AuthorizationBuilder</returns>
+        public ManagementBuilder WithDynamicDescriptor(string value) {
+            DynamicDescriptor = value;
             return this;
         }